requirements.txt
Pillow==9.0.1
PyMuPDF==1.19.6
